Private classes for CubeDomainsBuilder
