+++ core/src/main/scala/io/qbeast/core/model/CubeDomainsBuilder.scala
@@ -1,5 +1,6 @@
 package io.qbeast.core.model
 
+import io.qbeast.core.model.Weight.MaxValue
 import scala.collection.mutable
 
 object CubeDomainsBuilder {
@@ -128,7 +129,7 @@ class CubeDomainsBuilder protected (
     // Convert cube Weight and size into a Map[CubeId, (NormalizedWeight, Double)] from which
     // the cube domains are computed.
     weights.map { case (cubeId, weightAndCount) =>
-      cubeId -> weightAndCount.toWeightAndTreeSize(groupCubeSize)
+      cubeId -> weightAndCount.toWeightAndTreeSize(groupCubeSize, desiredCubeSize)
     }.toMap
 
   }
@@ -179,6 +180,153 @@ class CubeDomainsBuilder protected (
   }
 
 }
+
+private object WeightAndCountFactory {
+
+  def apply(
+      desiredCubeSize: Int,
+      groupCubeSize: Int,
+      cubeStatuses: Map[CubeId, CubeStatus]): WeightAndCountFactory = {
+    val isLeaf = (cubeId: CubeId) => !cubeId.children.exists(cubeStatuses.contains)
+    val multiplier = groupCubeSize / desiredCubeSize.toDouble
+
+    val cubeNormalizedWeights = cubeStatuses.map {
+      case (cubeId, status) if isLeaf(cubeId) =>
+        // TODO: Remove replicated data from cubeSize
+        val cubeSize = status.files.map(_.elementCount).sum
+        val groupSize = (cubeSize * multiplier).toLong.max(1L)
+        val nw = NormalizedWeight(groupCubeSize, groupSize).max(1.0)
+        cubeId -> nw
+      case (cubeId, status) => cubeId -> status.normalizedWeight
+    }
+
+    new WeightAndCountFactory(cubeNormalizedWeights, groupCubeSize)
+  }
+
+}
+
+/**
+ * Factory for WeightAndCount. The creation of which depends on whether the associated
+ * CubeId is present in the existing index, if it's a inner or leaf cube.
+ *
+ * @param existingNormalizedWeights Cube Weight of the existing index
+ * @param groupCubeSize partition-level cube size for indexing
+ */
+private class WeightAndCountFactory(
+    existingNormalizedWeights: Map[CubeId, NormalizedWeight],
+    groupCubeSize: Int) {
+
+  def createWeightAndCount(cubeId: CubeId): WeightAndCount =
+    existingNormalizedWeights.get(cubeId) match {
+      case Some(nw: NormalizedWeight) if nw < 1.0 =>
+        new InnerCubeWeightAndCount(Weight(nw))
+      case Some(nw: NormalizedWeight) =>
+        val existingLeafSize = ((groupCubeSize - 1) / nw).toInt
+        new LeafCubeWeightAndCount(existingLeafSize)
+      case None => new WeightAndCount(MaxValue, 0)
+    }
+
+}
+
+/**
+ * Weight and count for a CubeId.
+ * @param weight Weight of the associated CubeId
+ * @param count Element count of the associated CubeId
+ */
+private class WeightAndCount(var weight: Weight, var count: Int) {
+
+  /**
+   * Determines whether an instance, with its instanceWeight, should be included
+   * in the associated cube
+   * @param instanceWeight Weight of a given instance
+   * @param limit total capacity of the cube i.e. groupCubeSize
+   */
+  def shouldInclude(instanceWeight: Weight, limit: Int): Boolean =
+    count < limit
+
+  /**
+   * Update metadata to include the instance
+   * @param instanceWeight Weight of a given instance
+   */
+  def update(instanceWeight: Weight): Unit = {
+    count += 1
+    weight = instanceWeight
+  }
+
+  /**
+   * Convert to WeightAndTreeSize, with cube size i.e. count as the initial tree size value.
+   * @param gcs groupCubeSize
+   * @param dcs desiredCubeSize
+   */
+  def toWeightAndTreeSize(gcs: Int, dcs: Int): WeightAndTreeSize = {
+    val nw = toNormalizedWeight(gcs, dcs)
+    new WeightAndTreeSize(nw, cubeSize)
+  }
+
+  /**
+   * @return the number of added instances to the associated cube
+   */
+  def cubeSize: Int = count
+
+  /**
+   * Compute NormalizedWeight according to the state of the cube
+   * @param gcs groupCubeSize
+   * @param dcs desiredCubeSize
+   */
+  def toNormalizedWeight(gcs: Int, dcs: Int): NormalizedWeight =
+    if (count == gcs) NormalizedWeight(weight)
+    else NormalizedWeight(dcs, cubeSize)
+
+}
+
+/**
+ * WeightAndCount for an existing leaf cube.
+ * It can accept up to (groupCubeSize - start) records.
+ */
+private class LeafCubeWeightAndCount(start: Int) extends WeightAndCount(MaxValue, start) {
+
+  override def cubeSize: Int = count - start
+
+}
+
+/**
+ * WeightAndCount for an existing inner cube.
+ * It can accept up to groupCubeSize records, with the additional constraint
+ * of instanceWeight < existingWeight.
+ * @param existingWeight Weight for the associated existing CubeId
+ */
+private class InnerCubeWeightAndCount(existingWeight: Weight)
+    extends WeightAndCount(existingWeight, 0) {
+
+  override def shouldInclude(instanceWeight: Weight, limit: Int): Boolean = {
+    instanceWeight < existingWeight && count < limit
+  }
+
+  override def toNormalizedWeight(gcs: Int, dcs: Int): NormalizedWeight = {
+    // An existing inner cube should always remain as such
+    NormalizedWeight(weight)
+  }
+
+}
+
+/**
+ * Point, weight and parent cube identifier if available.
+ *
+ * @param point  the point
+ * @param weight the weight
+ * @param parent the parent
+ */
+private case class PointWeightAndParent(point: Point, weight: Weight, parent: Option[CubeId])
+
+/**
+ * NormalizedWeight and tree size of a given cube, with tree size defined as the
+ * sum of its own payload and all that of its descendents. The tree size is
+ * initialized with cube size and updated to the tree size during computation
+ * before being used.
+ * @param weight NormalizedWeight
+ * @param treeSize Cube tree size
+ */
+private class WeightAndTreeSize(val weight: NormalizedWeight, var treeSize: Double)
 
 /**
  * Cube bytes and its domain size from a given data partition, with domain defined as
