+++ src/test/scala/io/qbeast/spark/utils/QbeastTableTest.scala
@@ -1,7 +1,6 @@
 package io.qbeast.spark.utils
 
 import io.qbeast.TestClasses.Client3
-import io.qbeast.core.model.{CubeId, CubeStatus}
 import io.qbeast.spark.{QbeastIntegrationTestSpec, QbeastTable}
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.functions.col
